requirements.txt
flask
gunicorn
pillow
numpy
boto3
